Rename advanced deep link recipe
Refactored to `syntheticbackstack`. Changes include:

1. refactored package in main `app` from `advanced` to `syntheticbackstack` and updated README.md
2. renamed the deep link app from `advanceddeeplinkapp` to `syntheticbackstackapp`
3. refactored the packge name in `syntheticbackstackapp` from `advanced` to `syntheticbackstack`
4. updated the global README with new names and paths

Partially addresses #301
