--- advanceddeeplinkapp/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="app_name">Nav3 Deeplink App</string>
-</resources>
