+++ syntheticbackstackapp/src/main/java/com/example/nav3recipes/deeplink/syntheticbackstack/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.nav3recipes.deeplink.advanced.ui.theme
+package com.example.nav3recipes.deeplink.syntheticbackstack.ui.theme
 
 import androidx.compose.ui.graphics.Color
