+++ syntheticbackstackapp/src/main/res/drawable/outline_arrow_upward_24.xml
