+++ syntheticbackstackapp/src/main/res/values/colors.xml
